# docs/openapi-gpt-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "esdata API",
    "description": "API de consulta fiscal-regulatoria con trazabilidad a fuente oficial, superficies de governance y workers observables via sync_log.",
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "https://api.desuscribir.es",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/buscar": {
      "get": {
        "tags": [
          "buscar"
        ],
        "summary": "Buscar en legislacion consolidada",
        "operationId": "buscar",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Termino de busqueda",
              "title": "Q"
            },
            "description": "Termino de busqueda"
          },
          {
            "name": "fuente",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por fuente (boe, autonomica, etc.)",
              "title": "Fuente"
            },
            "description": "Filtrar por fuente (boe, autonomica, etc.)"
          },
          {
            "name": "ambito",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por ambito (tributario, etc.)",
              "title": "Ambito"
            },
            "description": "Filtrar por ambito (tributario, etc.)"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo de articulo",
              "title": "Tipo"
            },
            "description": "Filtrar por tipo de articulo"
          },
          {
            "name": "norma",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por codigo de norma (LIVA, LIRPF, etc.)",
              "title": "Norma"
            },
            "description": "Filtrar por codigo de norma (LIVA, LIRPF, etc.)"
          },
          {
            "name": "vigente_en",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Fecha de vigencia (YYYY-MM-DD)",
              "title": "Vigente En"
            },
            "description": "Fecha de vigencia (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegislacionSearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cnmv": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Listar Cnmv",
        "operationId": "listar_cnmv",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por texto o título",
              "title": "Q"
            },
            "description": "Filtrar por texto o título"
          },
          {
            "name": "ambito",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por ámbito regulatorio",
              "title": "Ambito"
            },
            "description": "Filtrar por ámbito regulatorio"
          },
          {
            "name": "tipo_documento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo de documento",
              "title": "Tipo Documento"
            },
            "description": "Filtrar por tipo de documento"
          },
          {
            "name": "vigencia",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por estado de vigencia",
              "title": "Vigencia"
            },
            "description": "Filtrar por estado de vigencia"
          },
          {
            "name": "regulacion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por regulación EU/ES relacionada (mifid_ii, mifir, mar, dora, priips, LIVMC)",
              "title": "Regulacion"
            },
            "description": "Filtrar por regulación EU/ES relacionada (mifid_ii, mifir, mar, dora, priips, LIVMC)"
          },
          {
            "name": "obligacion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo de obligación (presentacion_modelo, remision_informacion, control_interno, comunicacion_indicio, reporting_prudencial)",
              "title": "Obligacion"
            },
            "description": "Filtrar por tipo de obligación (presentacion_modelo, remision_informacion, control_interno, comunicacion_indicio, reporting_prudencial)"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset de paginación",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset de paginación"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Número de resultados (máx 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Número de resultados (máx 100)"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Campo de ordenación: fecha, referencia, titulo",
              "default": "fecha",
              "title": "Order By"
            },
            "description": "Campo de ordenación: fecha, referencia, titulo"
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Dirección de ordenación: asc, desc",
              "default": "desc",
              "title": "Order Dir"
            },
            "description": "Dirección de ordenación: asc, desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocInterpretativoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cnmv/buscar": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Buscar Cnmv",
        "operationId": "buscar_cnmv",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por texto o titulo",
              "title": "Q"
            },
            "description": "Filtrar por texto o titulo"
          },
          {
            "name": "ambito",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por ambito regulatorio",
              "title": "Ambito"
            },
            "description": "Filtrar por ambito regulatorio"
          },
          {
            "name": "tipo_documento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo de documento",
              "title": "Tipo Documento"
            },
            "description": "Filtrar por tipo de documento"
          },
          {
            "name": "vigencia",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por estado de vigencia",
              "title": "Vigencia"
            },
            "description": "Filtrar por estado de vigencia"
          },
          {
            "name": "regulacion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por regulacion EU/ES relacionada",
              "title": "Regulacion"
            },
            "description": "Filtrar por regulacion EU/ES relacionada"
          },
          {
            "name": "obligacion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo de obligacion",
              "title": "Obligacion"
            },
            "description": "Filtrar por tipo de obligacion"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset de paginacion",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset de paginacion"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Numero de resultados",
              "default": 20,
              "title": "Limit"
            },
            "description": "Numero de resultados"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Campo de ordenacion",
              "default": "fecha",
              "title": "Order By"
            },
            "description": "Campo de ordenacion"
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Direccion de ordenacion",
              "default": "desc",
              "title": "Order Dir"
            },
            "description": "Direccion de ordenacion"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocInterpretativoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cnmv/{referencia}/versions": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Get Cnmv Versions",
        "description": "Get version history for a CNMV document (Fase 23.6).",
        "operationId": "get_cnmv_versions",
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Referencia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CNMVVersionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cnmv/{referencia}/relaciones": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Get Cnmv Regulation Links",
        "description": "Get regulation links for a CNMV document (Fase 23.7).",
        "operationId": "get_cnmv_regulation_links",
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Referencia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CNMVRegulationLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cnmv/{referencia}/obligaciones": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Get Cnmv Obligation Links",
        "description": "Get obligation links for a CNMV document (Fase 23.9).",
        "operationId": "get_cnmv_obligation_links",
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Referencia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CNMVObligationLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cnmv/{referencia}": {
      "get": {
        "tags": [
          "cnmv"
        ],
        "summary": "Get Cnmv",
        "operationId": "get_cnmv",
        "parameters": [
          {
            "name": "referencia",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Referencia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocInterpretativoDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "CNMVObligationLinkResponse": {
        "properties": {
          "referencia": {
            "type": "string",
            "title": "Referencia"
          },
          "obligaciones": {
            "items": {
              "$ref": "#/components/schemas/CNMVObligationLinkItem"
            },
            "type": "array",
            "title": "Obligaciones"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "referencia",
          "obligaciones",
          "total"
        ],
        "title": "CNMVObligationLinkResponse"
      },
      "CNMVRegulationLinkItem": {
        "properties": {
          "regulacion_id": {
            "type": "string",
            "title": "Regulacion Id"
          },
          "relacion_tipo": {
            "type": "string",
            "title": "Relacion Tipo"
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          }
        },
        "type": "object",
        "required": [
          "regulacion_id",
          "relacion_tipo"
        ],
        "title": "CNMVRegulationLinkItem"
      },
      "CNMVRegulationLinkResponse": {
        "properties": {
          "referencia": {
            "type": "string",
            "title": "Referencia"
          },
          "regulaciones": {
            "items": {
              "$ref": "#/components/schemas/CNMVRegulationLinkItem"
            },
            "type": "array",
            "title": "Regulaciones"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "referencia",
          "regulaciones",
          "total"
        ],
        "title": "CNMVRegulationLinkResponse"
      },
      "CNMVVersionItem": {
        "properties": {
          "version_num": {
            "type": "integer",
            "title": "Version Num"
          },
          "cambio_tipo": {
            "type": "string",
            "title": "Cambio Tipo"
          },
          "fecha_version": {
            "type": "string",
            "title": "Fecha Version"
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          },
          "url_version": {
            "type": "string",
            "title": "Url Version"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          }
        },
        "type": "object",
        "required": [
          "version_num",
          "cambio_tipo",
          "texto"
        ],
        "title": "CNMVVersionItem"
      },
      "CNMVVersionResponse": {
        "properties": {
          "referencia": {
            "type": "string",
            "title": "Referencia"
          },
          "versiones": {
            "items": {
              "$ref": "#/components/schemas/CNMVVersionItem"
            },
            "type": "array",
            "title": "Versiones"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "referencia",
          "versiones",
          "total"
        ],
        "title": "CNMVVersionResponse"
      },
...
      "DocInterpretativoDetail": {
        "properties": {
          "referencia": {
            "type": "string",
            "title": "Referencia"
          },
          "fecha": {
            "type": "string",
            "title": "Fecha"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "tipo_documento": {
            "type": "string",
            "title": "Tipo Documento"
          },
          "ambito": {
            "type": "string",
            "title": "Ambito"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "url_fuente": {
            "type": "string",
            "title": "Url Fuente"
          },
          "url_aepd": {
            "type": "string",
            "title": "Url Aepd",
            "description": "Alias de URL fuente AEPD para consumidores MCP"
          },
          "organismo_emisor": {
            "type": "string",
            "title": "Organismo Emisor",
            "description": "Organismo emisor (cendoj)"
          },
          "estado_vigencia": {
            "type": "string",
            "title": "Estado Vigencia",
            "description": "Estado vigencia (cnmv)"
          },
          "numero_circular": {
            "type": "string",
            "title": "Numero Circular",
            "description": "Numero circular (cnmv)"
          },
          "fecha_publicacion": {
            "type": "string",
            "title": "Fecha Publicacion",
            "description": "Fecha publicacion (cnmv)"
          },
          "referencia_boe": {
            "type": "string",
            "title": "Referencia Boe",
            "description": "Referencia BOE (cnmv)"
          },
          "boe_referencia": {
            "type": "string",
            "title": "Boe Referencia",
            "description": "Alias de referencia BOE para consumidores MCP"
          },
          "url_cnmv": {
            "type": "string",
            "title": "Url Cnmv",
            "description": "Alias de URL fuente CNMV/BOE para consumidores MCP"
          }
        },
        "type": "object",
        "required": [
          "referencia",
          "tipo_documento",
          "ambito",
          "texto"
        ],
        "title": "DocInterpretativoDetail"
      },
...
      "DocInterpretativoListResponse": {
        "properties": {
          "documentos": {
            "items": {
              "$ref": "#/components/schemas/DocInterpretativoListItem"
            },
            "type": "array",
            "title": "Documentos"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocInterpretativoListItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Alias paginado para consumidores MCP/GPT"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "Offset paginacion (cnmv)"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Limite paginacion (cnmv)"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total resultados (cnmv)"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Offset paginacion"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Si hay mas resultados"
          },
          "next_offset": {
            "type": "integer",
            "title": "Next Offset",
            "description": "Siguiente offset"
          }
        },
        "type": "object",
        "required": [
          "documentos"
        ],
        "title": "DocInterpretativoListResponse"
      },
...
      "LegislacionSearchResponse": {
        "properties": {
          "q": {
            "type": "string",
            "title": "Q",
            "description": "Término de búsqueda"
          },
          "resultados": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Resultados"
          }
        },
        "type": "object",
        "required": [
          "q",
          "resultados"
        ],
        "title": "LegislacionSearchResponse"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Dedicated ESData API key for GPT Actions."
      }
    }
  }
}